+++ infra/etl/__init__.py
@@ -1,0 +1,10 @@
+"""ETL infrastructure module.
+
+Components:
+- ETL staging storage
+- Data Factory pipelines (optional)
+"""
+
+# TODO: Implement components
+# from .storage import EtlStorageComponent
+# from .data_factory import DataFactoryComponent