+++ infra/functions/__init__.py
@@ -1,0 +1,12 @@
+"""Azure Functions infrastructure module.
+
+Components:
+- Function App (workers)
+- App Service Plan (Consumption/Premium)
+- Function storage account
+"""
+
+# TODO: Implement components
+# from .function_app import FunctionAppComponent
+# from .app_service_plan import AppServicePlanComponent
+# from .storage import FunctionStorageComponent