+++ infra/gateway/__init__.py
@@ -1,0 +1,14 @@
+"""API Gateway infrastructure module.
+
+Components:
+- API Management instance
+- API definitions
+- Products & subscriptions
+- Policies (auth, rate limiting)
+"""
+
+# TODO: Implement components
+# from .apim import ApiManagementComponent
+# from .apis import ApiDefinition
+# from .products import ApiProduct
+# from .policies import AuthPolicy, RateLimitPolicy