+++ infra/messaging/__init__.py
@@ -1,0 +1,12 @@
+"""Messaging infrastructure module.
+
+Components:
+- Service Bus namespace
+- Queues
+- Topics & subscriptions
+"""
+
+# TODO: Implement components
+# from .servicebus import ServiceBusComponent
+# from .queues import QueueComponent
+# from .topics import TopicComponent